requirements.txt
pydantic[email]
fastapi
uvicorn[standard]
pymongo
motor
kafka-python
protobuf
googleapis-common-protos
passlib
python-jose
pycryptodome
aiokafka